move some logic so extended forms also correctly write instances
