+++ services/forms-from-config.ts
@@ -4,6 +4,7 @@ import formRepo from '../domain/data-access/form-repository';
 import comunicaRepo from '../domain/data-access/comunica-repository';
 import formExtRepo from '../domain/data-access/form-extension-repository';
 import { HttpError } from '../domain/http-error';
+import N3 from 'n3';
 
 const formsFromConfig: FormsFromConfig = {};
 const formDirectory = '/forms';
@@ -36,28 +37,63 @@ const fetchMetaTtlFromFormTtl = async (
   return await formRepo.getConceptSchemeTriples(conceptSchemeUris);
 };
 
+const mergeExtensionIntoBaseTtl = async (
+  baseFormTtl: string,
+  extensionFormTtl: string,
+): Promise<string> => {
+  const store = new N3.Store();
+  const mergeGraph = 'http://merge';
+
+  await formExtRepo.loadTtlIntoGraph(baseFormTtl, mergeGraph, store);
+  await formExtRepo.loadTtlIntoGraph(extensionFormTtl, mergeGraph, store);
+
+  await formExtRepo.deleteAllFromBaseForm(
+    ['form:targetType', 'form:targetLabel', 'ext:prefix', 'mu:uuid'],
+    mergeGraph,
+    store,
+  );
+
+  await formExtRepo.replaceFormUri(mergeGraph, store);
+  await formExtRepo.replaceExtendsGroup(mergeGraph, store);
+
+  return await formExtRepo.graphToTtl(mergeGraph, store);
+};
+
 export const fetchFormDefinitionById = async (
   formId: string,
 ): Promise<FormDefinition> => {
   const definitionFromConfig: FormDefinition | undefined =
     formsFromConfig[formId];
 
-  const formTtl = await computeIfAbsent(
+  let formTtl = await computeIfAbsent(
     definitionFromConfig || {},
     'formTtl',
     () => formRepo.fetchFormTtlById(formId),
   );
 
   if (!formTtl) throw new HttpError('Definition not found', 404);
+
+  if (await formExtRepo.isFormExtension(formTtl)) {
+    const extendedFormTtl = formTtl;
+    const baseFormUri = await formExtRepo.getBaseFormUri(formTtl);
+    const baseFormTtl = await fetchFormDefinitionByUri(baseFormUri);
+    if (!baseFormTtl) throw new HttpError('Definition not found', 404);
+    formTtl = await mergeExtensionIntoBaseTtl(
+      baseFormTtl.formTtl,
+      extendedFormTtl,
+    );
+  }
+
   if (!definitionFromConfig) formsFromConfig[formId] = { formTtl };
 
-  const metaTtl = await computeIfAbsent(definitionFromConfig, 'metaTtl', () =>
-    fetchMetaTtlFromFormTtl(formTtl),
-  );
+  // TODO this seems wrong, should metaTtl not combine all sources.
+  // const metaTtl = await computeIfAbsent(definitionFromConfig, 'metaTtl', () =>
+  //   fetchMetaTtlFromFormTtl(formTtl),
+  // );
 
   return {
     formTtl,
-    metaTtl,
+    metaTtl: null,
   };
 };
